requirements.txt
# Core dependencies
sentence-transformers>=2.2.0
torch>=1.9.0
transformers>=4.21.0

# Data processing
pandas>=1.3.0
numpy>=1.21.0

# System monitoring
psutil>=5.8.0

# Azure dependencies (if needed)
azure-search-documents>=11.4.0
azure-cosmos>=4.3.0

# Additional utilities
tqdm>=4.62.0
requests>=2.25.0
